Fix default values for getHistoryData
Since we write these times out to a packet, we need them to be actual Date objects or else things blow up. I could protect against writing null/undefined Dates as SLDateTimes, but I feel like an exception there is valid and warranted. I could set the default value in the message itself, but this feels like the most appropriate place to me. Honestly the failure to provide a from or a to (or if from isn't before to) should probably be an error itself, but I need to figure out the right way to pass that back.