Add weather forecast message
I was initially just trying to swallow the "forecast changed" event, and I assume it would bring the forecast with it, so I implemented parsing of the entire message. Then I discovered that "forecast changed" doesn't bring the forecast with it, so this turned out to be unnecessary. But I already did all the work, so might as well commit it. I don't think it's helpful, however, as my system is reporting forecast data for late last year and I'm not sure why. But this works (for that day's forecast...), so whatever.

I'm not bothering to document it since it currently seems useless. Maybe we'll figure it out later if anyone cares.