Fix handling of multiple messages in one packet
This is a pretty serious bug that can occur on low-powered hardware when multiple packets arrive quickly enough that they are passed along to the 'data' handler as one. With this fix, the packets are seen as separate and handled separately. Otherwise, we can end up never handing the message off to be processed because the buffer index never matched the expected message length; an ugly side effect of this problem is that the internal packet buffer would continue growing as long as the connection stayed active.