Improve FindUnits.search() to support retries more easily
Previously, every call to search() would bind() the socket and re-establish socket options in addition to actually sending the broadcast that looked for units. Now we bind the socket and set socket options once, then future calls to search() will simply issue another broadcast. This enables retries from consumer applications to work seamlessly. Documentation has been updated to ensure this is clearly explained.

I could probably change the bind() so that it happens in the constructor, but I didn't feel the need to upset the apple cart at this point since the library has been working this way since its inception.

I also updated the FindUnits object to only construct the message buffer once instead of every time we call sendServerBroadcast(). This is a bit of a micro-optimization, but there is just no reason to create this buffer every time we call the function, so it's an easy, harmless win.