Fix SLDateTime reading
I think this only fails on the last day of the month or during a certain window of time. date.setMonth() has an optional "day" argument that also fixed this problem, but I was nervous enough about the initial problem to go with the nuclear option of setting all date attributes at once so that the month/day/etc. resolution works properly.