SLUnit test passes, add back senderId
The previous method of setting sender id per UnitConnection didn't work out in practice since each Commands sub-object is itself an Outbound SLMessage which gets its senderId set at construction time which happens when the library first launches; any changes to the unit's sender id is not propagated down to these objects, so if each individual message doesn't give a sender id, it will use whatever was in place when the object was first constructed (which used to be random, but is now a predictable 0). Furthermore, the intent of the sender ID is to allow multiple calls to multiple APIs while retaining the ability to associate each response with a specific request, something that's much more complicated when the unit itself is the arbiter of the sender id.

Additionally, the sender ID wasn't being sent back with any responses since each response was its own object (as opposed to a child of Inbound which would have had the senderId property inherited automatically), so each of these has been updated to also package the senderId. Many portions of the API needed to be updated as a result, so a String wrapper, a Number wrapper, and a Bool wrapper exist for responses that don't need any other data (version, simple acks, etc.).

And miscellaneous fixes:

* closeAsync() was timing out and throwing an exception trying to remove a client if one had never been added; apparently the unit does not respond to removes if there's no matching client with that id.
* Re-instituted a method of initializing a UnitConnection from a LocalUnit response; this makes connecting to a client found with a local search dead-simple.
* Continued annotating parameter and return types, discovering a few bugs along the way.
* Fix a few "ack"-style responses that were always passing true instead of what the event emitted (previously the event also always emitted true, but any change to the event's response would not have been reflected).