Increase lint severity
Linebreaks I couldn't care less about; git will typically check out with the local linebreak style and commit with the remote style, so we can't possibly make this work everywhere (and it's fine so long as we're not mixing line endings).

Indent and quotes are both clear right now, so I'd prefer to error on them.