Enable noImplicitAny, fix fallout (#68)
Note: setEquipmentConfigurationAsync() was commented out because it was non functional as is. I have no idea what type the "data" parameter is supposed to be, but I have a feeling it was ported over from another codebase and not actually used here since I can't find any existing calls to it or existing types for it that would make sense.