Fix remote logins
There was a call to this.closeAsync() after connecting to the gateway which was attempting to resolve "this" as the socket object because of the function() being used instead of an arrow function.

A new Inbound() message was being processed with an attempt to fill in the controllerId and senderId of a "unit" that did not exist in the scope it was being called from; defaults are fine in this case.